requirements.txt
amqp==2.5.2
astroid==2.4.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
pylint==2.5.2
pytz==2020.1
redis==3.5.0
requests==2.23.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
